+++ src/index.web.tsx
@@ -1,0 +1,5 @@
+import Healthkit from './index.native'
+
+export * from './index.native'
+
+export default Healthkit