+++ src/utils/getWorkoutPlanById.ts
@@ -1,0 +1,7 @@
+import Native from '../native-types'
+
+async function getWorkoutPlanById(workoutUUID: string) {
+  return Native.getWorkoutPlanById(workoutUUID)
+}
+
+export default getWorkoutPlanById